Game logic skeleton
The game logic works. Movies can (in theory) send their ratings over to the game.js. The player can click on each button and have it compare each rating and determine if the player was correct or not. Scores and lives are updated (in logic). Removed old code from API.js and attempted to refactor the new code.