Game logic works fully
Game logic works fully, the player can play the game and it will loop until a game over correctly.

Added feedback to the top of the screen that tells the player when they can and cannot click the buttons. Also added a disclaimer stating that sometimes only one movie will generate.