Adjusted loading logic
Moved the loading logic from index.js to save.js. Adjusted the display of the saved data to use the table elements and rows. Also adjusted the sort function to sort by the score instead of the name of each element. Also added checks for reset state so the site will only look for those elements on game.html. Adjusted the load logic to only work on index.html